+++ mobile/android/android-components/components/feature/tabs/src/main/java/mozilla/components/feature/tabs/TabsUseCases.kt
@@ -5,10 +5,10 @@
 package mozilla.components.feature.tabs
 
 import mozilla.components.browser.session.Session
-import mozilla.components.browser.state.state.SessionState.Source
 import mozilla.components.browser.session.SessionManager
 import mozilla.components.browser.state.action.EngineAction
 import mozilla.components.browser.state.action.UndoAction
+import mozilla.components.browser.state.state.SessionState.Source
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.EngineSession.LoadUrlFlags
@@ -217,6 +217,20 @@ class TabsUseCases(
         }
     }
 
+    /**
+     * Use case for removing a list of tabs.
+     */
+    class RemoveTabsUseCase internal constructor(
+        private val sessionManager: SessionManager
+    ) {
+        /**
+         * Removes a specified list of tabs.
+         */
+        operator fun invoke(ids: List<String>) {
+            sessionManager.removeListOfSessions(ids)
+        }
+    }
+
     class RemoveAllTabsUseCase internal constructor(
         private val sessionManager: SessionManager
     ) {
@@ -272,6 +286,7 @@ class TabsUseCases(
     val addTab: AddNewTabUseCase by lazy { AddNewTabUseCase(store, sessionManager) }
     val addPrivateTab: AddNewPrivateTabUseCase by lazy { AddNewPrivateTabUseCase(store, sessionManager) }
     val removeAllTabs: RemoveAllTabsUseCase by lazy { RemoveAllTabsUseCase(sessionManager) }
+    val removeTabs: RemoveTabsUseCase by lazy { RemoveTabsUseCase(sessionManager) }
     val removeNormalTabs: RemoveNormalTabsUseCase by lazy { RemoveNormalTabsUseCase(sessionManager) }
     val removePrivateTabs: RemovePrivateTabsUseCase by lazy { RemovePrivateTabsUseCase(sessionManager) }
     val undo by lazy { UndoTabRemovalUseCase(store) }