+++ mobile/android/android-components/docs/changelog.md
@@ -12,6 +12,11 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/.config.yml)
 
+* **feature-tabs**
+  * Added `TabsUseCases.RemoveTabsUseCase` for removing an arbitrary list of tabs.
+* **browser-state**
+  * Added `TabListAction.RemoveTabsAction` to `BrowserAction`.
+
 * **feature-downloads**
   * 🚒 Bug fixed [issue #8823](https://github.com/mozilla-mobile/android-components/issues/8823) Downloads for data URLs were failing on nightly and beta more details in the [Fenix issue](https://github.com/mozilla-mobile/fenix/issues/16228#issuecomment-717976737).